Added some convenient functions and a "user define literal" to fractions.       bool isZero() const        { return _numerator == 0;      }       bool isNotZero() const     { return _numerator != 0;      }       bool gtZero() const        { return _numerator > 0;       }       bool geZero() const        { return _numerator >= 0;      }       bool ltZero() const        { return _numerator < 0;       }       bool leZero() const        { return _numerator <= 0;      }       bool isValid() const       { return _denominator != 0;    }
A Fraction constant can be written as (zero as example):
      0_Fr
      Fraction(0)
      Fraction(0,1)
