+++ libmscore/style.h
@@ -1151,6 +1151,7 @@ typedef std::vector<StyledProperty> ElementStyle;
 
 typedef std::array<StyledProperty, TEXT_STYLE_SIZE> TextStyle;
 
+enum class Tid;
 
 const TextStyle* textStyle(Tid);
 const TextStyle* textStyle(const char*);
